requirements.txt
flask>=2.3.0
firebase-admin>=6.2.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
ultralytics>=8.0.0
easyocr>=1.7.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
thefuzz>=0.19.0
pillow>=10.0.0
gunicorn>=21.2.0